Fix voting, admin commands and optional module support

+++ mods/Fifty.ServerMessages/mod/scripts/vscripts/fsm_killstreak.nut
@@ -11,28 +11,32 @@ array< int > killstreakAchivements = [ 100, 50, 40, 30, 20, 15, 10 ]
  * @param attacker The player who killed the victim
 */
 void function FSM_UpdateKillstreakInformation( entity victim, entity attacker ) {
-	if( attacker in killstreak ) {
-		killstreak[attacker]++
+	bool validAttacker = IsValid( attacker ) && attacker.IsPlayer() && attacker != victim
 
-		foreach( achivement in killstreakAchivements ) {
-			if( killstreak[attacker] == achivement ) {
-				bool usePopUp = GetConVarBool( "FSM_USE_RUI_POPUP_FOR_KILLSTREAK" )
-				string message
-				if( usePopUp )
-					message = attacker.GetPlayerName() + " is on a " + string( achivement ) + " kill streak!"
-				else
-					message = FSU_Highlight( attacker.GetPlayerName() ) + " is on a " + FSU_Highlight( string( achivement ) ) + " kill streak!"
-				
-				FSU_ChatBroadcast( message, usePopUp )
-				break
+	if( validAttacker ) {
+		if( attacker in killstreak ) {
+			killstreak[attacker]++
+
+			foreach( achivement in killstreakAchivements ) {
+				if( killstreak[attacker] == achivement ) {
+					bool usePopUp = GetConVarBool( "FSM_USE_RUI_POPUP_FOR_KILLSTREAK" )
+					string message
+					if( usePopUp )
+						message = attacker.GetPlayerName() + " is on a " + string( achivement ) + " kill streak!"
+					else
+						message = FSU_Highlight( attacker.GetPlayerName() ) + " is on a " + FSU_Highlight( string( achivement ) ) + " kill streak!"
+
+					FSU_ChatBroadcast( message, usePopUp )
+					break
+				}
 			}
+		} else {
+			killstreak[attacker] <- 1
 		}
-	} else {
-		killstreak[attacker] <- 1
 	}
 
 	if( victim in killstreak ) {
-		if( killstreak[victim] > 10 ) {
+		if( validAttacker && killstreak[victim] > 10 ) {
 			bool usePopUp = GetConVarBool( "FSM_USE_RUI_POPUP_FOR_KILLSTREAK" )
 			string message
 			if( usePopUp )
